initial merge static
