--- authentik/stages/authenticator/plugins/otp_static/__init__.py
@@ -1,4 +1,0 @@
-import django
-
-if django.VERSION < (3, 2):
-    default_app_config = "django_otp.plugins.otp_static.apps.DefaultConfig"
