--- authentik/stages/authenticator/plugins/otp_static/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class DefaultConfig(AppConfig):
-    name = "django_otp.plugins.otp_static"
-    default_auto_field = "django.db.models.AutoField"
