--- authentik/stages/authenticator_static/settings.py
@@ -1,5 +1,0 @@
-"""Static Authenticator settings"""
-
-INSTALLED_APPS = [
-    "django_otp.plugins.otp_static",
-]
